Use vmap for repeated mass profiles
This change avoids long compile times by removing the `for` loop from the code and replacing it with a `vmap` over the stacked inputs of the function.

In local testing with a model with 125 components this brought down the compile time for `ray_shooting` from about 40 seconds to 1 second.  The initialization for a Numpyro model went down from 6 mins to 30 seconds.

There might be scope to use `vmap` to also account for `alpha_static_single` at the same time, but this might be more hassle than it is worth as I think you would need to know how many keyword arguments each mass profile has before doing the `vmap`.